refactor(generator/dart): flatten the generator/dart directory structure a bit; rename google_cloud_common => google_cloud_gax
