+++ .github/workflows/generator_dart.yaml
@@ -59,17 +59,19 @@ jobs:
         working-directory: generator/dart
 
       # Analyze code.
-      - name: "analyze package:google_cloud_common"
-        run: dart analyze
-        working-directory: generator/dart/packages/google_cloud_common
-      - name: "analyze generated packages"
-        run: dart analyze
-        working-directory: generator/dart/packages/generated
+      - run: dart analyze examples
+        working-directory: generator/dart
+      - run: dart analyze generated
+        working-directory: generator/dart
+      - run: dart analyze packages
+        working-directory: generator/dart
+      - run: dart analyze tests
+        working-directory: generator/dart
 
       # Run tests.
-      - name: "test package:google_cloud_common"
+      - name: "test package:google_cloud_gax"
         run: dart test
-        working-directory: generator/dart/packages/google_cloud_common
+        working-directory: generator/dart/packages/google_cloud_gax
       - name: "test package:google_cloud_protobuf"
         run: dart test
-        working-directory: generator/dart/packages/generated/google_cloud_protobuf
+        working-directory: generator/dart/generated/google_cloud_protobuf
