+++ generator/dart/packages/google_cloud_gax/pubspec.yaml
@@ -11,17 +11,16 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-#
-# Code generated by sidekick. DO NOT EDIT.
 
-name: google_cloud_type
-description: The Google Cloud client library for the Common Types.
+name: google_cloud_gax
+description: Common code for the Google Cloud client library packages.
 version: 0.1.0
 
-environment: 
+environment:
   sdk: ^3.6.0
 
 resolution: workspace
 
 dependencies:
-  google_cloud_common: any
+  google_cloud_rpc: ^0.1.0
+  http: ^1.3.0
