+++ generator/dart/pubspec.yaml
@@ -19,8 +19,10 @@ environment:
   sdk: ^3.6.0
 
 workspace:
-  - packages/generated/google_cloud_protobuf
-  - packages/generated/google_cloud_rpc
-  - packages/generated/google_cloud_language
-  - packages/generated/google_cloud_type
-  - packages/google_cloud_common
+  - examples
+  - generated/google_cloud_language
+  - generated/google_cloud_protobuf
+  - generated/google_cloud_rpc
+  - generated/google_cloud_type
+  - packages/google_cloud_gax
+  - tests
