+++ generator/dart/tests/pubspec.yaml
@@ -1,0 +1,14 @@
+name: tests
+description: Unit and integration tests for the generated cloud APIs.
+
+publish_to: none
+
+environment: 
+  sdk: ^3.6.0
+
+resolution: workspace
+
+dev_dependencies:
+  google_cloud_protobuf: any
+  google_cloud_rpc: any
+  test: ^1.25.0
