+++ generator/testdata/dart/protobuf/golden/secretmanager/lib/secretmanager.dart
@@ -24,8 +24,8 @@ library;
 
 import 'dart:typed_data';
 
-import 'package:google_cloud_common/common.dart';
-import 'package:google_cloud_common/src/json_helpers.dart';
+import 'package:google_cloud_gax/common.dart';
+import 'package:google_cloud_gax/src/json_helpers.dart';
 import 'package:google_cloud_protobuf/protobuf.dart';
 import 'package:http/http.dart' as http;
