+++ generator/testdata/dart/protobuf/golden/secretmanager/pubspec.yaml
@@ -24,6 +24,6 @@ environment:
 resolution: workspace
 
 dependencies:
-  google_cloud_common: any
+  google_cloud_gax: any
   google_cloud_protobuf: any
   http: any
